+++ controlplane/CHANGELOG.md
@@ -4,6 +4,16 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.213.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.213.0...controlplane@0.213.1) (2026-04-02)
+
+**Note:** Version bump only for package controlplane
+
+# [0.213.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.212.0...controlplane@0.213.0) (2026-04-01)
+
+### Features
+
+* pass `proxy url` to `admission webhook` ([#2718](https://github.com/wundergraph/cosmo/issues/2718)) ([bcb4ed3](https://github.com/wundergraph/cosmo/commit/bcb4ed34ed78bfa5fad072a7ebf62ed648d7d872)) (@wilsonrivera)
+
 # [0.212.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.211.0...controlplane@0.212.0) (2026-03-31)
 
 ### Bug Fixes